refactor: view, domain 패키지 추가 및 이동
