+++ src/main/view/Message.java
@@ -1,4 +1,4 @@
-package main;
+package main.view;
 
 public enum Message {
